Add a flag, --suggest-classpaths, that suggests jars to add to the classpath if Vogar fails due to missing classes. It does this by building a map of classes and packages to the jars in which they are contained, and then when it sees a failure like "package foo.bar does not exist" in the outcome it tells you a jar that you can add to the classpath to get it.
To find the jars, it looks at the directories in a colon-separated environment variable called VOGAR_JAR_PATH. I have mine set to <DALVIK ROOT>/out/target/common/obj/JAVA_LIBRARIES/, since this seems to be where most of the important jars are kept. Right now it only looks for jars called "classes.jar", but it's easy to extend it to look for other types of jars as well.

It takes a few seconds to build the index the first time you use it, but retrieves a cached version very quickly on subsequent runs. It rebuilds the cache daily unless you manually delete it (from ~/.vogar/classfileindex).

Example:

If you run:

vogar --suggest-classpaths <DALVIK ROOT>/libcore/luni/src/test/java/tests/sql/StatementTest.java

it will spit out:

"Warning: consider adding the following to the classpath:
<DALVIK ROOT>/out/target/common/obj/JAVA_LIBRARIES/sqlite-jdbc_intermediates/classes.jar"

I don't know how generally useful this is, but it's saved me quite a bit of time when trying to get tests to run properly.

git-svn-id: http://vogar.googlecode.com/svn/trunk@85 aa685c63-decc-881d-cd2b-7fa72aad72e1