Have vogar automatically pull XML files that it finds in the temporary directory after the test/benchmark has terminated. This is mainly to retrieve Caliper failure output.
The device-version is a bit lamer (no directory recursing, except a special-cased "caliper-results") since unlike on the host, there isn't currently a great way of checking if a given file is a directory or not (that I know of). This could be rememdied by improving androidSdk.ls() to check file metadata.

git-svn-id: http://vogar.googlecode.com/svn/trunk@137 aa685c63-decc-881d-cd2b-7fa72aad72e1