Change how host + target processes communicate.
It used to read XML over a socket. This was robust but omitted native output to stdout and stderr.

Now it reads JSON over stdout. The socket code is still in place (but not currently used) because this change breaks activity mode. I'll do that as a follow up commit.

git-svn-id: http://vogar.googlecode.com/svn/trunk@178 aa685c63-decc-881d-cd2b-7fa72aad72e1