Automatically add libcore resource directories to the runtime classpath.
Note that I added a new flavor of classpath. You might think I could have
just reused the existing --classpath, but that's supplied at compile time
as well as run time, and there might be all kinds of confusing files in
the resource tree that shouldn't be shown to the compiler. (This is certainly
the case with libcore. You can't build with the resource directories on the
compile time classpath.)



git-svn-id: http://vogar.googlecode.com/svn/trunk@285 aa685c63-decc-881d-cd2b-7fa72aad72e1