Add basic JUnit4 support
This adds basic support for JUnit4 test cases. It can do all the basic
assertions and also supports the "RunWith" annotation with the
"Parameterized" and "Suite" runners.

It doesn't properly implement BeforeClass and AfterClass yet. Instead
it runs those once for each test. This can be improved in the future,
but should run correctly until then.

This was written without looking at the JUnit source code so it may
have compatibility bugs.


git-svn-id: http://vogar.googlecode.com/svn/trunk@290 aa685c63-decc-881d-cd2b-7fa72aad72e1