Disable JCA security checks on the RI
This allows the running of tests that insert their own Provider class. This
was failing silently before, but replacing this one field with a Map that
returns true bypasses the checks.

This should probably log an error when it can't find the class via reflection
on platforms where it expects to find it.


git-svn-id: http://vogar.googlecode.com/svn/trunk@295 aa685c63-decc-881d-cd2b-7fa72aad72e1