Update vogar's caliper version to fix a bug
This change contains the tip of the v0.5-rc1 branch from:
https://code.google.com/p/caliper/source/browse/?name=v0.5-rc1
and so constitutes a minor upgrade. Since vogar last
took a caliper snapshot the gson library has been split out.

Because that branch is old I have so far been unable to upstream
fix for Android in that branch. The fix is:

Since the L release, Android returns a special sub-class of
Properties from System.getProperties(). The Properties instance
falls back to a set of immutable properties for certain platform
values. If treated as a Map, i.e. using Map methods like get()
instead of getProperty(), it will appear empty.

This change copies the Properties object to produce a Map instead,
which has all the expected values in it.

The fix is already in the caliper HEAD on googlecode.

---
 caliper/src/main/java/com/google/caliper/EnvironmentGetter.java | 4 ++--
 1 file changed, 2 insertions(+), 2 deletions(-)

Bug: https://code.google.com/p/android/issues/detail?id=81363
Change-Id: Id5c38dfc729d581ec1addeddc2ee8ca70d34625b